Properly replace constant effects on tuple properties
The replacement machinery allows an effect whose value won't change any
more to be replaced by a ConstantEffect. This allows that effect (and
anything it refers to) to be garbage collected.

Allow this to work with tuple properties; specifically, make
_TupleExtractEffect and _TupleMakeEffect replace themselves correctly,
have _TupleElementProperty delegate its do_replacements to its parent
TupleProperty, and do replacements after an tuple element is __set__ or
__delete__d.

Also, change the names of _TupleMakeEffect's parents so that the one named
"parent" is the one more suitable when following properties' parent
chains, i.e. the full property.

A couple of new tests is included.