Make Clock.advance() stop on scheduled event
Instead of keeping remaining advancement time separately,
schedule an event at the appropriate time and then just check if
it has elapsed.
This event is special in that it is scheduled *after* other
events of the same time.

Allow advance() to be called with arbitrary futures, or with
None which will "run forever" (as long as events are still
scheduled).

Add tests for the new functionality.

Also, improve documentation.